+ ajustes pequenos de CSS, Edicao 100% dos blocos de conteudo

+++ app/Http/Controllers/ProfissionalController.php
@@ -139,7 +139,9 @@ class ProfissionalController extends AppBaseController
         $profissional = $this->profissionalRepository->update($request->all(), $id);
 
         if ($request->file) {
-            $profissional->fotoListagem->delete();
+            if ($profissional->fotoListagem()->first()) {
+                $profissional->fotoListagem->delete();
+            }
 
             $foto = $this->fotoRepository->uploadAndCreate($request);
             $profissional->fotoListagem()->save($foto);
